content/media/video/test/test_wav_trunc_seek.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: seeking in truncated file</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<audio id='v'
       onloadedmetadata='return startTest();'
       onseeked='return endSeek();'
       onended='return endTest();'>
  <source type='audio/x-wav' src='r11025_s16_c1_trunc.wav'>
</audio>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if the ended event works correctly.
var v = document.getElementById('v');
var endPassed = false;
var completed = false;

function startTest() {
  if (completed)
    return false;

  ok(v.duration > 0.3 && v.duration < 0.5, "Duration should be around 0.4: " + v.duration);

  v.currentTime = 1.0;
  return false;
}

function endTest() {
  ok(v.ended, "Checking ended set after seeking to EOF and playing");
  SimpleTest.finish();
  return false;
}

function endSeek() {
  if (completed)
    return false

  completed = true;
  ok(v.currentTime >= 0.3 && v.currentTime <= 0.5,
     "Checking currentTime at end: " + v.currentTime);
  v.play();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
